Data Detail: Create resources to movie detail's data layer

+++ core/src/main/kotlin/br/com/mxel/cuedot/domain/detail/GetMovieDetail.kt
@@ -8,12 +8,12 @@ class GetMovieDetail(
         private val repository: IMovieDetailRepository
 ) {
 
-    fun execute(movieId: Int): Observable<Event<Movie>> {
+    fun execute(movieId: Long): Observable<Event<Movie>> {
 
         return Observable.concat(Observable.just(Event.loading()), getMovieDetail(movieId))
     }
 
-    private fun getMovieDetail(movieId: Int): Observable<Event<Movie>> {
+    private fun getMovieDetail(movieId: Long): Observable<Event<Movie>> {
 
         return if (movieId < 1) {
             Observable.just(Event.error(DetailError.INVALID_ID))
